requirements.txt
kafka-python
psycopg2
requests
